+++ src/vmm/src/vmm_config/drive.rs
@@ -51,6 +51,8 @@ pub struct BlockDeviceConfig {
     /// If set to true, the drive is opened in read-only mode. Otherwise, the
     /// drive is opened as read-write.
     pub is_read_only: Option<bool>,
+    /// If set to true, the drive advertises discard support to the guest.
+    pub discard: Option<bool>,
     /// Path of the drive.
     pub path_on_host: Option<String>,
     /// Rate Limiter for I/O operations.
@@ -208,6 +210,7 @@ mod tests {
                 partuuid: self.partuuid.clone(),
                 is_root_device: self.is_root_device,
                 is_read_only: self.is_read_only,
+                discard: self.discard,
                 cache_type: self.cache_type,
 
                 path_on_host: self.path_on_host.clone(),
@@ -237,6 +240,7 @@ mod tests {
             cache_type: CacheType::Writeback,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path),
             rate_limiter: None,
             file_engine_type: None,
@@ -271,6 +275,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(true),
+            discard: None,
             path_on_host: Some(dummy_path),
             rate_limiter: None,
             file_engine_type: None,
@@ -303,6 +308,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(true),
+            discard: None,
             path_on_host: Some(dummy_path),
             rate_limiter: None,
             file_engine_type: None,
@@ -332,6 +338,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_1),
             rate_limiter: None,
             file_engine_type: None,
@@ -348,6 +355,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_2),
             rate_limiter: None,
             file_engine_type: None,
@@ -375,6 +383,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_1),
             rate_limiter: None,
             file_engine_type: None,
@@ -391,6 +400,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_2),
             rate_limiter: None,
             file_engine_type: None,
@@ -407,6 +417,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_3),
             rate_limiter: None,
             file_engine_type: None,
@@ -448,6 +459,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_1),
             rate_limiter: None,
             file_engine_type: None,
@@ -464,6 +476,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_2),
             rate_limiter: None,
             file_engine_type: None,
@@ -480,6 +493,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_3),
             rate_limiter: None,
             file_engine_type: None,
@@ -522,6 +536,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_1.clone()),
             rate_limiter: None,
             file_engine_type: None,
@@ -538,6 +553,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_2.clone()),
             rate_limiter: None,
             file_engine_type: None,
@@ -610,6 +626,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_1),
             rate_limiter: None,
             file_engine_type: None,
@@ -626,6 +643,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(false),
+            discard: None,
             path_on_host: Some(dummy_path_2),
             rate_limiter: None,
             file_engine_type: None,
@@ -652,6 +670,7 @@ mod tests {
             cache_type: CacheType::Unsafe,
 
             is_read_only: Some(true),
+            discard: None,
             path_on_host: Some(dummy_file.as_path().to_str().unwrap().to_string()),
             rate_limiter: None,
             file_engine_type: Some(FileEngineType::Sync),
@@ -682,6 +701,7 @@ mod tests {
             cache_type: CacheType::default(),
 
             is_read_only: Some(true),
+            discard: None,
             path_on_host: Some(backing_file.as_path().to_str().unwrap().to_string()),
             rate_limiter: None,
             file_engine_type: None,